+++ server/src/types/responseType.ts
